lexen-legal-ai/pyproject.toml
[tool.poetry]
name = "lexen-legal-ai"
version = "0.1.0"
description = "Lexen Legal AI Assistant using CrewAI and Ollama"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = {extras = ["tools"], version = "^0.70.1"}
python-dotenv = "^1.0.0"
litellm = "^1.44.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
